training/requirements.txt
# Core dependencies
numpy>=1.24.0
rlgym>=1.2.0
stable-baselines3>=2.0.0
torch>=2.0.0

# Data processing
pandas>=2.0.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Config
pyyaml>=6.0.0